move EntityInteractData into own class to fix PMD CouplingBetweenObjects
