+++ src/test/java/org/betonquest/betonquest/quest/objective/interact/EntityInteractObjectiveTest.java
@@ -20,31 +20,31 @@ class EntityInteractObjectiveTest {
 
     @Test
     void testCreateFromMinimalCountingDataWithoutInteractAppendix() throws QuestException {
-        final EntityInteractObjective.EntityInteractData data = new EntityInteractObjective.EntityInteractData("1", null, null);
+        final EntityInteractData data = new EntityInteractData("1", null, null);
         assertEquals(NO_PROGRESS_ON_SIMPLE_INTERACT_SERIALIZED, data.toString(), "Data should serialize from default data.");
     }
 
     @Test
     void testCreateFromExtendedCountingDataWithoutInteractAppendix() throws QuestException {
-        final EntityInteractObjective.EntityInteractData data = new EntityInteractObjective.EntityInteractData("1/1/1/0", null, null);
+        final EntityInteractData data = new EntityInteractData("1/1/1/0", null, null);
         assertEquals(NO_PROGRESS_ON_SIMPLE_INTERACT_SERIALIZED, data.toString(), "Data should serialize from serialized data without UUID list.");
     }
 
     @Test
     void testCreateFromMinimalCountingDataWithInteractAppendix() throws QuestException {
-        final EntityInteractObjective.EntityInteractData data = new EntityInteractObjective.EntityInteractData("1;", null, null);
+        final EntityInteractData data = new EntityInteractData("1;", null, null);
         assertEquals(NO_PROGRESS_ON_SIMPLE_INTERACT_SERIALIZED, data.toString(), "Data should serialize from minimal data with empty UUID list.");
     }
 
     @Test
     void testCreateFromExtendedCountingDataWithInteractAppendix() throws QuestException {
-        final EntityInteractObjective.EntityInteractData data = new EntityInteractObjective.EntityInteractData(NO_PROGRESS_ON_SIMPLE_INTERACT_SERIALIZED, null, null);
+        final EntityInteractData data = new EntityInteractData(NO_PROGRESS_ON_SIMPLE_INTERACT_SERIALIZED, null, null);
         assertEquals(NO_PROGRESS_ON_SIMPLE_INTERACT_SERIALIZED, data.toString(), "Data should serialize from serialized data.");
     }
 
     @Test
     void testCreateDataWithCollected() throws QuestException {
-        final EntityInteractObjective.EntityInteractData data = new EntityInteractObjective.EntityInteractData("2/1/1/1;00000000-0000-0000-0000-000000000000", null, null);
+        final EntityInteractData data = new EntityInteractData("2/1/1/1;00000000-0000-0000-0000-000000000000", null, null);
         assertEquals("2/1/1/1;00000000-0000-0000-0000-000000000000", data.toString(), "Data should serialize with UUID.");
     }
 }
